Bug 1830914 - Use 'defaults_fn' for custom parameter defaults
This is the more official way of setting up parameter defaults and
solves a couple edge cases.
